Try this:
1. Create a new archive in ark and call it something like 'test.gz' or
   'test.bz2'
2. Add a file to it

Notice that the filename shown is NOT the name of the file you just added.
It will be 'test' (or whatever the basename of your archive is).

This makes sense programatically since gunzipping test.gz leaves you with
a file called test... but what a nightmare from a newbie UI standpoint!

This is a quick patch from Xandros to just pop up a dialog explaining what
is going to happen:

 You are creating a simple compressed archive which contains only one input file.
 When uncompressed, the file name will be based on the name of the archive file.
 If you add more files you will be prompted to convert it to a real archive.

svn path=/trunk/kdeutils/ark/; revision=166131