Support for 7-Zip archives.
FEATURE: 48267

In order to deal with 7-Zip archives using Ark, one needs to install p7zip from http://p7zip.sf.net. This needs testing. Ark CVS HEAD is compatible with KDE libs 3.3, so you need not to install an unstable KDE to help in testing. If you can help, please contact me for details.

Please note that adding files to 7-Zip archives seems to be really slow (though much of this slowness might be due to the fact that my only pc is an old K6-2, but anyway it seems to be slower than other archive formats), but nothing can be done on Ark's side to improve this.

Known bugs:
	Timestamps show the wrong year
	Some errors do not trigger Error MessageBoxes.

CCMAIL: kde-quality@kde.org

svn path=/trunk/kdeutils/ark/; revision=355145