Backport r1003290.
Strip "./" from entries being deleted.

The code to strip the leading "./" from entries has been moved to a new
private method, so methods for adding and removing data can use it.
Currently, the "./" was being stripped only when adding entries, which
resulted in the view not being correctly updated when an entry was
removed.

CCBUG: 200173

svn path=/branches/KDE/4.3/kdeutils/ark/; revision=1003295