Backport r1004414.
Save and remember settings in the extraction dialog.

The code to load and remember the 'open destination' and 'preserve
paths' has been moved to ExtractionDialog itself.

A new <label> tag has been included in ark.kcfgc. I don't know if this
is a string freeze violation (in this case, to fix a bug), so I'm ccing
kde-i18n-doc.

CCBUG: 201854
CCMAIL: kde-i18n-doc@kde.org

svn path=/branches/KDE/4.3/kdeutils/ark/; revision=1004429