Relicense most of kerfuffle under the 2-clause BSD license.
The initial Kerfuffle code by Henrique Pinto was already BSD-licensed, and Harald and I have agreed to relicense our bits under the same license.

This is library code, and we prefer to use a BSD-style license here.

There are two files that need to be relicensed, though: kerfuffle_export.h and observer.h. Will contact the authors shortly.

svn path=/trunk/KDE/kdeutils/ark/; revision=1067630