Backport r1119906.
When opening an HTML file with KHTMLPart, disable Java, JavaScript,
plugins and external references.

This is a saner and safer default, since the archive might come from an
unknown and untrusted sender. If access to the disabled stuff is needed,
the user should simply extract the file and view it in a browser.

See also bug 235468.

CCBUG: 235546

svn path=/branches/KDE/4.4/kdeutils/ark/; revision=1119908