Backport r1124326.
Escape file names in CliInterface when necessary.

The infrastructure has been added, and is used by the clizip plugin so
that it is now possible to extract and delete files which have
characters such as '[' or '*' correctly (infozip accepts some wild
characters to make our lives little bit more difficult).

BUG: 208091

svn path=/branches/KDE/4.4/kdeutils/ark/; revision=1124327