Add JSONArchiveInterface, a parser for test data written in JSON.
JSONArchiveInterface is more complete compared to TextArchiveInterface,
since we can easily set other properties for each entry besides
FileName.

Currently, the JSON format accepted by the parser is the following:

{
    "entries": [
        { ... },
        { ... },
    ]
}

Where ... is a series of key-value pairs where the key is a value from
Kerfuffle::EntryMetaDataType. An example archive will be committed
shortly.

svn path=/trunk/KDE/kdeutils/ark/; revision=1137693