Backport r1140173.
Revamp the clirar plugin.

This was prompted by bug 242071: RAR files have this concept of
subheaders, which include comments, NTFS streams and other things I have
no idea since the format is undocumented.

Subheaders can be ignored during the listing, but they confused the old
parser code because they were unexpected and broke the expected listing
pattern.

In order to fix that, we now call unrar with 'vt' instead of just 'v' so
that we get more information. However, this means more lines are output
for each entry, so our "parser" needed to be improved.

The code looks much cleaner and (hopefully) easier to understand now.

Furthermore, the old code probably ignored files whose names started
with '*'. Finally, some TODOs have been added for some aspects we should
watch out when dealing with the code.

CCBUG: 242071

svn path=/branches/KDE/4.4/kdeutils/ark/; revision=1140180