Backport r1142819.
Do not show the Compressed Size column for 7z archives.

7z archives apparently store only the compressed size for the whole file,
so the values would not be shown correctly in that column.

We now store the archive type when parsing the output from 7z, since the
compressed size values work fine with other archive types 7z is able to
read, such as zip or bzip2.

BUG: 236696
FIXED-IN: 4.4.5

svn path=/branches/KDE/4.4/kdeutils/ark/; revision=1142821