Make it possible to load the extraction dialog from the menu.
Before the Quick Extraction list, we add an action that launches the
Extraction dialog and a separator. It should now be possible to launch
the extraction dialog by both clicking the "Extract" button on the
toolbar and via the menu (actually, it is possible to launch it by
choosing "Extract To" in the pop-up menu in the toolbar, but that is a
minor side-effect we can live with).

Based on a patch by Simon COURTOIS, thanks a lot!

BUG: 190507
CCMAIL: scourtois@linagora.com
FIXED-IN: 4.6.0

svn path=/trunk/KDE/kdeutils/ark/; revision=1184983