Add the CPIO mimetypes to the list of mimetypes Ark supports.
In order to do that, kerfuffle_libarchive.desktop has been split into
two desktop files.

One of them, kerfuffle_libarchive.desktop, contains a MimeType entry
with all the mimetypes which we can read _and_ write. The other one,
kerfuffle_libarchive_readonly.desktop, contains the list of mimetypes
we can open read-only.

For now, we also move the mimetypes for Debian packages and ISO files
to the list of mimetypes we can only read. Debian packages are
actually CPIO archives, so it should be moved back to
kerfuffle_libarchive.desktop when we implement write support for CPIO
archives in general.

Thanks to George Metaxas for bringing this issue up again :)

BUG: 160875

svn path=/trunk/KDE/kdeutils/ark/; revision=1194909