Rework clirarplugin's line reading logic again.
The code looks more like a state machine, which is good and should make it
easier to understand.

This also made it possible to add some code to check the column names written
before the header lines and decide whether we're using unrar-free or unrar
based on that (which should fix Ark crashing when being used with unrar-free).

If we're using unrar-free, we never try to parse the third file name entry
line, as unrar-free never outputs it.

It should be hopefully safe to backport, but I'll wait at least a few days. As
I'm unsure whether it's OK to backport it, the FIXED-IN version will be 4.6.0.

BUG: 243273
FIXED-IN: 4.6.0


svn path=/trunk/KDE/kdeutils/ark/; revision=1200794