Remove the temporarily extracted previewed file correctly.
After r1067160, the QDialog::exec() call in ArkViewer was replaced by
a call to QDialog::show().

This, in turn, made the call to KIO::NetAccess::del() be reached
before the dialog was closed.

We now manually remove the previewed file in all possible places in
ArkViewer::view() with QFile::remove(), and call KRun::runUrl() with
tempFile=true so that kioexec also takes care of removing the
previewed file if that's the case.

Finally, when ArkViewer::dialogClosed() is reached, it also tries to
remove the file (this is the expected code path).

Unfortunately, the code got uglier, so in the future we should try to
centralize the calls to QFile::remove() in a single place.

BUG: 261785
FIXED-IN: 4.5.5

svn path=/branches/KDE/4.5/kdeutils/ark/; revision=1210723