Backport r1211296.
Check for the header string when entering the entry details line too.

The rar file might be corrupted, so we may get a 'Unexpected end of archive'
line instead of a file name, and the next line is the header string instead
of the entry's details.

Thanks to Nicolas L?\195?\169cureuil for pointing that out on IRC.

svn path=/branches/KDE/4.5/kdeutils/ark/; revision=1211297