cliinterface: Merge create the process creation and execution methods.
Inpired by what Okular does in its comicbook generator, it also makes
sense for us to remove the process we run in the same method (ie
executeProcess()).

findProgramAndCreateProcess, createProcess and executeProcess have
been merged: the process is now crated, run and destroyed in the same
method, which makes us not need to wory about deleting the process in
different places.

While at it, doKill()'s implementation could also be simplified, as we
waitForFinished() right after calling KProcess::start(), so we need
not call it a few more times in doKill().

svn path=/trunk/KDE/kdeutils/ark/; revision=1216557