libarchive: Consider renamed files valid files when extracting.
When one or more files are selected for extraction and they already
exist on the filesystem, if the user chooses to rename it, extraction
will fail for that file. That happens because when we are not
extracting all files, the new file name fails the
files.contains(entryName) check.

We now keep track of the current file being renamed to avoid this
problem.

BUG: 266159
FIXED-IN: 4.6.1


svn path=/branches/KDE/4.6/kdeutils/ark/; revision=1220439