libarchive: Consider renamed files valid files when extracting.
When one or more files are selected for extraction and they already
exist on the filesystem, if the user chooses to rename it, extraction
will fail for that file. That happens because when we are not
extracting all files, the new file name fails the
files.contains(entryName) check.

We now keep track of the current file being renamed to avoid this
problem.

Forward-port of r1220439.

CCBUG: 266159

svn path=/trunk/KDE/kdeutils/ark/; revision=1220441