AddDialog: Add a default file extension to the file being created.
KFileDialog (actually, KFileWidget)'s does not append an extension to the
selected file when it already exists, which is always the case when a single
file is being compressed, as the default file name is the file's own name.

To work around this, we always add the main extension for the current mimetype
the file passed to setSelection().

CCBUG: 272914

svn path=/trunk/KDE/kdeutils/ark/; revision=1231314