Do not pass widget(), which is a QSplitter, as a parent to KFileDialog.
Oh, we all love nested event loops. A QSplitter is passed as a parent to
KFileDialog, which creates a KFileWidget. KFileWidget runs a nested event loop
to stat the specified directory, which leads to posted events being processed,
including QSplitter::childEvent(). childEvent() calls KFileDialog::show(),
which shows the KFileDialog in a non-modal state. When KFileDialog::exec() is
called, the dialog is already shown and is non-modal.

BUG: 264819
FIXED-IN: 4.6.4

svn path=/branches/KDE/4.6/kdeutils/ark/; revision=1231646