cliinterface: Do not reset the password after a wrong one was entered.
This partially reverts r964726. There does not seem to be any solid
reason for resetting the password like this, and there is at least one
reason not to do so: when in batch-extraction mode, addExtraction()
will create a ListJob to check if an archive has a single folder. If
the archive is header-encrypted, it will ask for a password. In case a
wrong password is entered, preserving it means the coming
ExtractionJob will use the wrong password and finally show the
"Incorredt password" error message which was not shown before (and
cancel the forthcoming jobs).

This should fix the issue in comment #75 of bug 35371.

svn path=/trunk/KDE/kdeutils/ark/; revision=1233141