Backport r1216556, r1216557, r1216558, r1216559, r1216762, r1216763, r1224452, r1233140 and r1233141 (!!).
This should make password support in header-protected archives work again.

As support for password-protected archives with unencrypted headers
was already present, this should finish minimally resolving bug 35371
once more. More work on user-visible messages (status messags, error
messages etc) is probably needed, but not essential.

BUG: 35371
FIXED-IN: 4.6.4

--

r1216556 by rkcosta:

cliinterface: Move the qRegisterMetaType code to the constructor.

It is going to be run only once anyway, so it does not make much sense
to run the check in every process call.

--

r1216557 by rkcosta:

cliinterface: Merge create the process creation and execution methods.

Inpired by what Okular does in its comicbook generator, it also makes
sense for us to remove the process we run in the same method (ie
executeProcess()).

findProgramAndCreateProcess, createProcess and executeProcess have
been merged: the process is now crated, run and destroyed in the same
method, which makes us not need to wory about deleting the process in
different places.

While at it, doKill()'s implementation could also be simplified, as we
waitForFinished() right after calling KProcess::start(), so we need
not call it a few more times in doKill().

--

r1216558 by rkcosta:

cliinterface: Use KPtyProcess on non-Windows platforms.

Once again inspired by Okular's comicbook generator (and also by Ark's
code circa the 4.2 release, which also seems to have borrowed the same
ideas from Okular), we now use a KPtyProcess on non-Windows platforms.

Some programs (such as unrar) open a TTY in occasions such as showing
password prompts, as they usually use getpass(3).

This is the first step towards supporting header-protected rar and 7z
files.

--

r1216559 by rkcosta:

cliinterface: Add the PasswordPromptPattern parameter.

Plugins can now specify a regular expression that matches password
prompts, which are now detected by CliInterface.

Header-protected rar and 7z files should be working now.

--

r1216762 by rkcosta:

archivebase: Run user queries on the internal ListJob.

After the recent commits, a password may be prompted on the archive's first
listing (listIfNotListed()) if the archive is header-protected.

We now process this occasional user query, so Ark does not get stuck when
extracting header-protected files in batch extraction mode.

--

r1216763 by rkcosta:

cliinterface: Remove m_program, as it is not needed anymore.

--

r1224452 by rkcosta:

cliinterface: Set the correct boolean value after QEventLoop::exec().

QEventLoop::exec() does not return an int, but a bool, and 0 means it
ran successfully -- however, it will translate into false after the
implicit cast, which is the inverse of what we want.

--

r1233140 by rkcosta:

cliinterface: Wait a few seconds before killing the process.

In some cases, failOperation() will try to kill an application which
is already in the process of finishing gracefully, so it is better to
wait a few seconds before trying to kill the process.

--

r1233141 by rkcosta:

cliinterface: Do not reset the password after a wrong one was entered.

This partially reverts r964726. There does not seem to be any solid
reason for resetting the password like this, and there is at least one
reason not to do so: when in batch-extraction mode, addExtraction()
will create a ListJob to check if an archive has a single folder. If
the archive is header-encrypted, it will ask for a password. In case a
wrong password is entered, preserving it means the coming
ExtractionJob will use the wrong password and finally show the
"Incorredt password" error message which was not shown before (and
cancel the forthcoming jobs).

This should fix the issue in comment #75 of bug 35371.

svn path=/branches/KDE/4.6/kdeutils/ark/; revision=1234597