cliinterface: Do not assume a process does not exist in runProcess().
The Q_ASSERT(!m_process) in the beginning of runProcess() was wrong, as when
one is adding files to an archive (or creating an archive) processFinished()
calls list(), which in turn calls runProcess() as well, thus making the assert
fail.

The best thing to do is just wait for the existing process to finish before
launching the new one.

BUG: 277393
FIXED-IN: 4.7.0

svn path=/branches/KDE/4.7/kdeutils/ark/; revision=1240907