Do not connect to KJob::result() from the worker thread.
This is another step towards getting rid of the threading mess we
currently have.

We previously connected to KJob::result() from the worker thread using
Qt::DirectConnection just to make sure QThread::quit() was called
right away (synchronously). We now make the connection from the main
thread without Qt::DirectConnection, and just make sure we start the
worker thread's event loop if there is an actual need to.

The other use of Qt::DirectConnection in the worker thread (via
QMetaObject::invkeMethod) has also been removed, as it was there to
make sure the method was called right away from the worker thread as
well.

svn path=/trunk/KDE/kdeutils/ark/; revision=1240914