Move the JSON parsing code to a separate class.
Besides splitting the code a little bit and having some typedefs that make the
code cleaner, this is a first step towards being able to use the parser in
other classes (this is particularly useful once there are tests for
cliinterface).

Additionally, we stop compiling jsonarchiveinterface.cpp with every test, as it
is stupid. Instead, we now create a library and link to it.

svn path=/trunk/KDE/kdeutils/ark/; revision=1240916