Move destination directory retrieval to ExtractJob.
Part::m_destinationDirectory was only set when the files were
extracted using the extraction dialog.

This meant that if the user had chosen to open the destination
directory after an extraction he would always have his home directory
open when the extraction was done via quick extract or drag'n'drop, as
KRun::runUrl would have an empty KUrl as its parameter.

We now make the destination directory passed to ExtractJob publicly
available, and retrieve it from there.

Forwardport of r1242112.

CCBUG: 188857

svn path=/trunk/KDE/kdeutils/ark/; revision=1242114