Specify when the extraction dialog's settings should be followed.
On ExtractJob's side, we make the extraction options it received
publicly accessible.

On Part's side, we set a "FollowExtractionDialogSettings" option to
true only when the extraction dialog was used to extract files, and
only actually follow these settings (for now, open the destination
directory and/or quit Ark after the extraction) when that option is
set.

Forwardport of r1242113.

CCBUG: 188857

svn path=/trunk/KDE/kdeutils/ark/; revision=1242115