clizip: Detect whether an entry is a directory in a dumber way.
infozip seems to have some trouble indentifying whether some entries
are directories in certain zip files, and outputs the wrong
information in the entry's attributes.

Simply checking if an entry's name ends in '/' seems to be more
reliable.

This also uncovered a crash in Dolphin, so I'm not closing the bug
right now.

The Ark side of the bug should be fixed in 4.7.1.

CCBUG: 280354