cliinterface: Fix order of FileExistInput options in the apidox.
The cliinterface code and the plugins use 3 for autoskip and 4 for
cancel, so make the documentation match the actual behavior.

This is part of review request 102702 by Theofilos Intzoglou, but I
found it better to have it in a separate commit.

CCMAIL: int.teo@gmail.com
(cherry picked from commit cd7336de390062b01b4963dadfd7b6699409926e)