Kill process when no cancel option exists for a cli tool.
Some cli programs (such as zipinfo) do not have an option to cancel the
extraction when a file already exists. In this case, we resorted to
autoskipping, which is not really the same thing.

When a program does not have the cancel option, we now resort to closing
it manually.

Patch by Theofilos Intzoglou, thanks a lot!

REVIEW: 102702