Restore the ability to use Ark as an embedded KPart
To support that, automatically generate the MIME type list
in the application, part and Konqueror plugin desktop files.
This keeps the long list of MIME types consistent and up-to-date.

Each plugin declares the MIME types that it supports in its
CMakeLists.txt file, either fixed, or depending on configuration.
This is inserted into the plugin's own desktop file, and the overall
list for all plugins is inserted into the top level desktop files.

Explicitly check for libarchive RPM support, rather than looking
for a specific library version.  The test that we really want is
for a library version greater or equal, but CMake doesn't support
that directly (and CMake 2.6 not easily).

REVIEW:102803

Original review at http://svn.reviewboard.kde.org/r/5405/