Move Kerfuffle::factory() into Archive itself.
The factory() function was used to create an Archive instance, which was
never created directly anywhere. Instead, make Archive's constructor
private and provide static create() methods which do exactly what the
factory() function did.

We now accept a `parent' argument, which should help fix some leaks in
BatchExtract and AddToArchive. The leaks didn't help when ArkPart was
used, as ArchiveModel was responsible for manually deleting the Archive
instance in its destructor (doing so is not needed now that we
explicitly set the model as the archive's parent).