cliinterface: A password prompt can also be shown when extracting.
This is a condition that is not currently exercised, but trying to fix
bug 291282 in another way made this problem more evident.

Suppose a user is asked for the archive password before the CLI
program is actually called to extract some files. If she presses OK
without entering a password, the code in CliInterface::copyFiles()
would not pass $PasswordSwitch to the CLI program because the password
is empty. In the case of infozip, it would then prompt the user for a
password, which would leave Ark stalled because there was no check for
a password prompt while already running the CLI program in extraction
mode.