Install a hicolor version of Ark's icon.
In a similar fashion to Okular commit df5364d (and probably others
spread across KDE's repositories), we need to install a hicolor
version of our icons so things such as menu entries are shown with an
actual icon in non-KDE desktops.

... And it is time for some history digging once again.

Until the 3.5 series, Ark had its set of icons which were correctly
installed according to the icon naming spec. During the transition
from the Crystal to the Oxygen icon set, for some reason it was
decided that some applications should start using more generic names
for their icons. Some of those names are defined in the icon naming
freedesktop.org spec, so other desktops are expected to have their own
versions available anyway. KCalc uses accessories-calculator, for
example.

According to [1], however, some icons fell in the category of

  "not in the spec, but let's keep it out of there anyways,
   it's not really suited to be a standard icon".

This happened to "utilities-file-archiver", which was added to the
Oxygen icon set and is used by a few different KDE applications.

This in itself would be fine, as Oxygen is a dependency for KDE
applications anyway, so KIconLoader and friends end up always finding
it.

However, if one is using another desktop, the algorithm it uses to
find icons for the applications shown in a panel or something like
that is usually the one defined in the icon theme XDG spec. Which
means Oxygen is normally not looked for, so the application gets no
icon (or has a default icon).

Git commit b479aea4ede1fb99886eb345e8f1fa80d9a6a354 then made Ark and
its desktop files use "utilities-file-archiver" in the Icon entry, and
commit bc40140adff29cba8af1ad4ee5039eef5c63fef0 moved the existing
icons out of Ark, since they were part of the Crystal icon set.

We are now sort of reverting those commits by copying the
"utilities-file-archiver" icons from Oxygen into Ark
again. infopanel.cpp still refers to "utilities-file-archiver" due to
the explanation about KIconLoader given above.

[1] http://techbase.kde.org/Projects/Oxygen/namingSpec/apps

BUG: 287547
FIXED-IN: 4.8.5