Stop emitting finished() from CliInterface::failOperation().
We were always emitting finished() twice if failOperation() was
called:
 - If it did kill the process in its doKill() call, the fact that the
   process was killed would call processFinished(), which would emit
   finished(true). failOperation() would later emit finished(false).
 - For the cases in which the process was not being killed, the
   callers in cliinterface.cpp were acting according to the following
   pattern:

       if (something failed) {
           emit error("Oops, something bad happened");
           failOperation();
           return false;
       }

   If the process wasn't killed it means it had already finished, so
   processFinished() had already been called (and consequently
   finished(true) had already been emitted), yet failOperation()
   would emit finished(false) again.

This was particularly troublesome for AddToArchive and BatchExtract,
since they rely on the KJob::result() signal to kill the
AddToArchive/BatchExtract instance and quit the application. Since
our Jobs connect to the finished() signal to call KJob::emitResult(),
the application was being deleted by the time the second finished()
signal was emitted, leading to crashes (this normally happens if you
enter a wrong password or extract a corrupted file).

Unfortunately, we still don't unit test CliInterface, so no there are
no new tests :(

BUG: 304764
BUG: 304178
FIXED-IN: 4.9.1