Delete m_process earlier.
Deleting and setting m_process to 0 at the end of runProcess() is
sometimes too late.

If one connects to KJob::result() to delete an archive interface, that
is all going to happen when finished() is emitted in processFinished()
and before `delete m_process' is called in runProcess(). This ends up
leaving us with a non-NULL m_process, causing the assert in
CliInterface's destructor to fail.

Delete m_process and set it to 0 in processFinished() itself, since it
is not supposed to be used afterwards anyway.

BUG: 304067
BUG: 301278
FIXED-IN: 4.9.1