Reintroduce the different code path for setting a directory icon.
Reintroduce the call removed in 5e9bfc33861c91a0db604f4a0969c9d15bad318a,
but in a different way.

Some tarballs (such as the one I have for asciidoc 8.6.4) only report files,
so the parent directories need to be created manually. However, this means
we pass a path such as "foo", without a trailing slash, to
KMimeType::findByPath(), which then does not determine that that should have
a "inode/directory" mime type.

We now pass this mime type directly if the ArchiveEntry passed to setEntry()
is a directory.