libarchive: Artificially add a trailing slash for rootNodes.
Some tarballs (such as the one I have for asciidoc 8.6.4) only report the
files in it, so the directories need to be created manually by Ark. However,
this means the ArchiveEntry created in ArchiveModel::parentFor() will not
have a trailing slash, and so will the path passed as RootNode to
LibArchiveInterface::copyFiles(). The truncation that happens later will
then leave a slash at the beginning if we're not extracting a top-level
entry, which will simply make drag'n'drop extraction fail.