clirar: Update comment about service headers.
It turns out there is some documentation on the RAR file format in
http://www.rarlab.com/technote.htm, so it is not true that the format is
undocumented.

While here, use the term "service header" instead of "subheader" as in the
file format documentation. It would make sense to rename the enum we use as
well, but it would clobber the history and git blame without much benefit.