rar: Do not crash when unrar lists a symlink.
RAR v3 and v4 output an extra line when an entry is a symlink. This line
contains the symlink target. Since we were not accounting for this extra
line before, our parser crashed.

Work around the issue by checking if we're listing a symlink and skipping an
extra line if so.

BUG:		331065
FIXED-IN:	4.12.3