cli7zip: Change the precedence order of the 7z binaries we look for.
Instead of going from least capable (7zr) to most capable (7z), do the
opposite. Looking for 7zr first means that even though a 7z binary capable
of handling a certain archive type might be present on the system it will
not be used. This is particularly troublesome because 7zr does not support
archives with password.

This should not cause any problems for distros that ship only some 7zip
binaries since we still look for all 3 of them (7z, 7za and 7zr).

BUG:		329941
FIXED-IN:	4.12.4