Revert "Allow multiple selection in archiveview"
The commit being reverted seems to work fine for the ZIP and LHA
backends, but for libarchive and RAR it can extract files with the wrong
name, a varying hierarchy depending on which of the selected entries is
being dragged or sometimes even leave selected entries out.

While it is certainly bad to go back to the previous state, it is worse
to have this working at varying degrees in each backend, the worst case
including partial extraction with wrong names.

This reverts commit 52ac4aea6f391cfe43dec0994891095d39364bd2.

CCBUG: 187152