Reset info panel and window caption when an archive fails to open.
When opening an archive that for whatever reason fails to load (because
it's corrupt, or an invalid format) the archive appears to be open when
it is not in fact open. This patch fixes that by resetting the open
archive, info panel name and window caption.

Patch by Mathias Tillman <master.homer@gmail.com>, thank you!

BUG:      245705
CCMAIl:   master.homer@gmail.com
FIXED-IN: 4.14.1
REVIEW:   120027