wip: Allow multiple entries to be drag'n'dropped.
This is the beginning of the work towards fixing bugs 208384 and 187152.

Group selected items by closest unselected parent and create a separate
ExtractJob for each of them, as this is the only way to make sure
unnecessary parent directories are not created.

ZIP files still create the whole directory structure, but that's always
been the case (due to info zip itself) and can be fixed separately.

TODO:
- Selected directories are not pulling its child files and directories.
- The ExtractJobs must be run sequentially, as parallel jobs crash Ark
  when a CliInterface-based plugin (zip, rar etc) is being used.