Switch from libarchive2 to libarchive3.
Ark now depends on libarchive3.
Symbols deprecated in libarchive2 and removed in libarchive3 has been replaced,
as well as the symbols deprecated in libarchive3 which will be removed in libarchive4.
At the same time, we can now remove the #ifdef checks in the libarchive plugin,
since those symbols are always defined in libarchive3.
The FoundLibArchive cmake module has been removed, since it's now built-in in the newer
CMake versions.

REVIEW: 121342