Fix password never being asked again after a wrong one has been entered
Previously, if the user entered a wrong password for a password-
protected archive, Ark would remember that a wrong password had been
entered and had to be closed down to allow entering the password again.

This commit fixes this issue by presenting the "Enter password" dialog
again when the user attempts some action that requires access to the
password-protected files.
BUG: 297820
FIXED-IN: 15.07.80
REVIEW: 123358