Implement a cancelled() signal for CliInterface
Add a cancelled() signal for ReadOnlyArchiveInterface and implement it
for CliInterface. Previously, when the user tried to preview a file in a
password-protected zip archive and then cancelled the password dialog,
the ExtractJob would not be given a finished() signal and SetReadyGui()
would not be triggered in the KPart leading to the UI appearing to be
loading indefinitely. Now, finished() is called for the ExtractJob when
the password dialog is cancelled but cancelled() is used to signal to
the KPart not to open a viewer in this case.