Only open preview when clicking files if CTRL or SHIFT is not pressed
Add a slotClicked() slot and connect it to the clicked() signal of the
QTreeView. The slot only calls slotPreviewWithInternalViewer() if CTRL
or SHIFT is not pressed simultaneously. This fixes selection of multiple
files and also enables deselection of a file by holding CTRL and
clicking it. This is more consistent with the usual behavior of desktop
applications.
BUG: 344640
FIXED-IN: 15.07.80