Improve hide/show logic for information panel
The logic governing hiding/showing of the infopanel as well as
storing/retrieving the size of the QSplitter was reworked into a more
intuitive and simple approach. A configuration option "showInfoPanel" of
type bool was added, which stores whether or not the infopanel should be
shown. The option "splitterSizesWithBothWidgets" was removed, since it's
no longer needed. Default values of "splitterSizes" are now set in the
kcfg file, instead of in part/part.cpp.

REVIEW: 123980